Some code cleanup around the S3 build cache provider.
